Feature dev

+++ src/config/Http.ts
@@ -1,6 +1,5 @@
 import axios from "axios";
 const baseURL = `${process.env.NEXT_PUBLIC_PYTHON_BASE_URL}`;
-
 const Http = () => {
   const instance = axios.create({ baseURL, withCredentials: true });
   instance.interceptors.request.use(async (request) => {
@@ -18,22 +17,29 @@ const Http = () => {
     async (error) => {
       if (error.response) {
         const { status } = error.response;
+        const originalRequest = error.config;
 
         if (status === 401) {
-          localStorage.clear();
-          if (localStorage.getItem("access_token")) {
-            const data = await refreshToken();
-            if (data) {
-              return await axios.request(error.config);
+          if(!originalRequest._retry){
+            originalRequest._retry = true;
+           if (localStorage.getItem("access_token")) {
+              const data = await refreshToken();
+              if (data.success) {
+                instance.defaults.headers.common[
+                  "Authorization"
+                ] = `Bearer ${data.access}`;
+                return await instance(error.config);
+              } else {
+                localStorage.clear();
+                window.location.href = "/login";
+              }
             } else {
               localStorage.clear();
-              window.location.href = "/login";
+              return error?.response || { error: "An error occurred" };
             }
-          } else {
-            return error?.response || { error: "An error occurred" };
           }
         } else if (status === 500) {
-          // window.location.href = "/server-error";
+          window.location.href = "/server-error";
         }
       }
       return error?.response || { error: "An error occurred" };
@@ -47,14 +53,13 @@ export default Http();
 
 const refreshToken = async () => {
   const instance = axios.create({ baseURL, withCredentials: true });
-  const refresh=localStorage.getItem("refresh_token")
-  const { data } = await instance.post("user/refresh-token",{refresh});
-  
-  if (data && data.access && data.refresh) {
+  const refresh = localStorage.getItem("refresh_token");
+  const { data } = await instance.post("user/refresh-token/", { refresh });
+  if (data && data?.access && data?.refresh) {
     localStorage.setItem("access_token", data.access);
     localStorage.setItem("refresh_token", data.refresh);
     axios.defaults.headers.common["Authorization"] = `Bearer ${data.access}`;
-    return true;
+    return { success: true, access: data.access };
   }
-  return false;
+  return { success: false };
 };
